refactor: extract `AddResourcePopover`
